+++ spike/version_rev.py
@@ -10,8 +10,8 @@ except:
     revision = "--not determined--"
 ProgramName = 'SPIKE'
 VersionName = 'Development version'
-version = '0.99.2'
-rev_date = '06-03-2019'
+version = '0.99.3'
+rev_date = '11-03-2019'
 
 def report():
     "prints version name when SPIKE starts"